Move glue.qt.widgets.glue_mdi_area to glue.app.qt.mdi_area
